+++ services/index.js
@@ -1,5 +1,5 @@
 module.exports = {
 	...require('./auth'),
 	...require('./movies'),
-	...require('./watchParties')
+	...require('./watchParty')
 };
